requirements.txt
python-telegram-bot==20.7
web3==6.11.3
solana==0.30.2
ccxt==4.1.13
ta==0.10.2
python-dotenv==1.0.0 